Refactor Electron main into modular files
Split main process responsibilities into separate modules and simplify main.mjs: create electron/windows.mjs (window creation), electron/preflight.mjs (preflight/update/CEA startup checks), and electron/ipc.mjs (IPC handlers). Update main.mjs to use startSplash/startMain flows, call runPreflightChecks, register IPC handlers, and keep killCEAProcess lifecycle handling. This improves separation of concerns and makes startup logic and window creation easier to test and maintain.
